Ensure validation of all PreferenceForm fields
Add question and employee fields of Preference Model as PreferenceForm fileds. This ensures field validation and prevents attempts for insert null values into to the database.

This, however does not solve the root cause of the problem - the fact that frontend can control the formset parameters and force django into creation of new model instances. I don't *think* there is a solution for that in formset itself so the best course of action would be to create a separate class / decorator for formset classes that are meant **only** to modify existing data. This would require an override of functions that are responsible for instance creation like `save_new_objects` and additionally the number of returned forms could be compared against the size of the queryset used for initial population of the formset.